[DX] Move MongoDB NormalizedData normalizers
