+++ src/Pim/Bundle/CatalogBundle/MongoDB/Normalizer/NormalizedData/MetricNormalizer.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\MongoDB\Normalizer;
+namespace Pim\Bundle\CatalogBundle\MongoDB\Normalizer\NormalizedData;
 
-use Pim\Component\Catalog\Model\ProductPriceInterface;
+use Pim\Component\Catalog\Model\MetricInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
- * Normalize a product price collection to store it as mongodb_json
+ * Normalize a product metric to store it as mongodb_json
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductPriceNormalizer implements NormalizerInterface
+class MetricNormalizer implements NormalizerInterface
 {
     /**
      * {@inheritdoc}
@@ -23,7 +23,10 @@ class ProductPriceNormalizer implements NormalizerInterface
         if ($object->getData() !== null) {
             $data = [
                 'data'     => $object->getData(),
-                'currency' => $object->getCurrency()
+                'unit'     => $object->getUnit(),
+                'baseData' => $object->getBaseData(),
+                'baseUnit' => $object->getBaseUnit(),
+                'family'   => $object->getFamily()
             ];
         }
 
@@ -35,6 +38,6 @@ class ProductPriceNormalizer implements NormalizerInterface
      */
     public function supportsNormalization($data, $format = null)
     {
-        return $data instanceof ProductPriceInterface && 'mongodb_json' === $format;
+        return $data instanceof MetricInterface && 'mongodb_json' === $format;
     }
 }
